Escape special characters in consoleUri
When we log the console uri for a failed work item, the uri is not escaped, despite the fact that we save the metadata with the escaped path. This leads to urls that are not navigable. This change moves the uri path construction to a helper function, and calls that function from where we originally construct the uri and where we log it.